scripts/requirements.txt
beautifulsoup4==4.12.3
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.32
langchain_openai==0.1.21
langchain_text_splitters==0.2.2
python-dotenv==1.0.1
requests==2.28.1
faiss-cpu==1.8.0